chess-openings-tierlist/requirements.txt
# Core dependencies
berserk>=0.14.0
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.4.0
python-dateutil>=2.8.2

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Database (sqlite3 is built into Python)

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0

# Development
pytest>=7.4.0
black>=23.0.0